user

+++ src/main/java/tn/esprit/piproject/Repositories/RoleRepository.java
@@ -1,11 +1,12 @@
 package tn.esprit.piproject.Repositories;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import tn.esprit.piproject.Entities.Role;
+
 import tn.esprit.piproject.Entities.ERole;
+import tn.esprit.piproject.Entities.Role;
 
 import java.util.Optional;
 
-public interface RoleRepository extends MongoRepository<Role ,Long> {
-    Optional<Role> findByName(ERole name);
+public interface RoleRepository extends MongoRepository<Role, Integer> {
+	  Optional<Role> findByName(ERole name);
 }
